Tests/LibWeb/Text/expected/abspos-relayout-shadow-dom.txt

shadow-child=55,35,30,20
partial-layouts=0 full-layouts=1
FAIL
